+++ packages/components/Avatar/src/Avatar.types.ts
@@ -1,9 +1,9 @@
 import type { IViewProps } from '@fluentui-react-native/adapters';
-import { ImageProps, ViewProps, TextProps, ColorValue } from 'react-native';
-import { IBackgroundColorTokens, IForegroundColorTokens, IBorderTokens, FontTokens } from '@fluentui-react-native/tokens';
-import { PresenceBadgeProps, BadgeSize, PresenceBadgeStatus } from '@fluentui-react-native/badge';
-import { IconProps, IconSourcesType } from '@fluentui-react-native/icon';
-import { SvgProps } from 'react-native-svg';
+import type { ImageProps, ViewProps, TextProps, ColorValue } from 'react-native';
+import type { IBackgroundColorTokens, IForegroundColorTokens, IBorderTokens, FontTokens } from '@fluentui-react-native/tokens';
+import type { PresenceBadgeProps, BadgeSize, PresenceBadgeStatus } from '@fluentui-react-native/badge';
+import type { IconProps, IconSourcesType } from '@fluentui-react-native/icon';
+import type { SvgProps } from 'react-native-svg';
 
 export const AvatarName = 'Avatar';
 export const AvatarSizesForTokens = [