+++ packages/components/Button/src/ToggleButton/ToggleButton.tsx
@@ -2,10 +2,12 @@
 import * as React from 'react';
 import { Pressable } from 'react-native';
 import { ActivityIndicator } from '@fluentui-react-native/experimental-activity-indicator';
-import { ToggleButtonProps, toggleButtonName, ToggleButtonType } from './ToggleButton.types';
+import type { ToggleButtonProps, ToggleButtonType } from './ToggleButton.types';
+import { toggleButtonName } from './ToggleButton.types';
 import { TextV1 as Text } from '@fluentui-react-native/text';
 import { stylingSettings } from './ToggleButton.styling';
-import { compose, mergeProps, withSlots, UseSlots } from '@fluentui-react-native/framework';
+import type { UseSlots } from '@fluentui-react-native/framework';
+import { compose, mergeProps, withSlots } from '@fluentui-react-native/framework';
 import { Icon, createIconProps } from '@fluentui-react-native/icon';
 import { buttonLookup } from '../Button';
 import { useToggleButton } from './useToggleButton';