+++ packages/components/Menu/src/MenuPopover/MenuPopover.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
-import { compressible, mergeProps, patchTokens, useFluentTheme, UseTokens } from '@fluentui-react-native/framework';
+import type { UseTokens } from '@fluentui-react-native/framework';
+import { compressible, mergeProps, patchTokens, useFluentTheme } from '@fluentui-react-native/framework';
 import { Callout } from '@fluentui-react-native/callout';
-import { menuPopoverName, MenuPopoverProps, MenuPopoverTokens } from './MenuPopover.types';
+import type { MenuPopoverProps, MenuPopoverTokens } from './MenuPopover.types';
+import { menuPopoverName } from './MenuPopover.types';
 import { useMenuPopover } from './useMenuPopover';
 import { useMenuPopoverTokens } from './MenuPopoverTokens';
 import { Platform, View } from 'react-native';