+++ packages/experimental/Avatar/src/NativeAvatar.tsx
@@ -1,6 +1,8 @@
 /** @jsx withSlots */
-import { compose, UseSlots, buildProps, mergeProps, withSlots } from '@fluentui-react-native/framework';
-import { ImageURISource, NativeModules, ViewProps, ColorValue } from 'react-native';
+import type { UseSlots } from '@fluentui-react-native/framework';
+import { compose, buildProps, mergeProps, withSlots } from '@fluentui-react-native/framework';
+import type { ImageURISource, ViewProps, ColorValue } from 'react-native';
+import { NativeModules } from 'react-native';
 import { ensureNativeComponent } from '@fluentui-react-native/component-cache';
 
 const avatarName = 'NativeAvatar';