+++ packages/experimental/Tabs/src/TabsItemTokens.macos.ts
@@ -1,6 +1,6 @@
-import { TabsItemTokens } from './TabsItem.types';
-import { TokenSettings } from '@fluentui-react-native/use-styling';
-import { Theme } from '@fluentui-react-native/framework';
+import type { TabsItemTokens } from './TabsItem.types';
+import type { TokenSettings } from '@fluentui-react-native/use-styling';
+import type { Theme } from '@fluentui-react-native/framework';
 
 export const tabsItemStates: (keyof TabsItemTokens)[] = ['hovered', 'selected', 'focused', 'disabled'];