+++ packages/framework/framework/src/compressible.test.tsx
@@ -1,9 +1,11 @@
 /** @jsx withSlots */
-import { buildUseTokens, UseTokens } from './useTokens';
+import type { UseTokens } from './useTokens';
+import { buildUseTokens } from './useTokens';
 import { applyTokenLayers } from '@fluentui-react-native/use-tokens';
-import { TextProps, TextStyle, Text, View } from 'react-native';
+import type { TextProps, TextStyle } from 'react-native';
+import { Text, View } from 'react-native';
 import { compressible } from './compressible';
-import { Theme } from '@fluentui-react-native/theme-types';
+import type { Theme } from '@fluentui-react-native/theme-types';
 import { mergeStyles } from '@fluentui-react-native/merge-props';
 import { useSlot, withSlots } from '@fluentui-react-native/use-slot';
 import * as React from 'react';