+++ packages/theming/apple-theme/src/__tests__/apple-theme.test.ts
@@ -1,6 +1,6 @@
 import { createMacOSColorAliasTokens, createMacOSShadowAliasTokens } from '../createMacOSAliasTokens';
 import { getIsHighContrast, setIsHighContrast } from '../appleHighContrast.macos';
-import { AppearanceOptions } from '@fluentui-react-native/theme-types';
+import type { AppearanceOptions } from '@fluentui-react-native/theme-types';
 import { createAppleTheme } from '../createAppleTheme';
 
 // For some reason the automatically mocked AccessibilityInfo object is the react-native version, not the