+++ packages/components/Callout/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @fluentui-react-native/callout
 
-This log was last generated on Mon, 30 Jan 2023 07:12:38 GMT and should not be manually modified.
+This log was last generated on Fri, 03 Feb 2023 01:49:36 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.21.59
+
+Fri, 03 Feb 2023 01:49:36 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.17
+- Bump @uifabricshared/foundation-compose to v1.13.0
+
+## 0.21.58
+
+Thu, 02 Feb 2023 01:29:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.16
+- Bump @uifabricshared/foundation-compose to v1.12.39
+
+## 0.21.57
+
+Wed, 01 Feb 2023 22:33:15 GMT
+
+### Patches
+
+- Enable @typescript-eslint/consistent-type-imports (30809111+acoates-ms@users.noreply.github.com)
+- Bump @fluentui-react-native/adapters to v0.10.1
+- Bump @fluentui-react-native/component-cache to v1.4.3
+- Bump @fluentui-react-native/interactive-hooks to v0.22.15
+- Bump @fluentui-react-native/tokens to v0.20.7
+- Bump @uifabricshared/foundation-composable to v0.11.2
+- Bump @uifabricshared/foundation-compose to v1.12.38
+- Bump @uifabricshared/foundation-settings to v0.12.2
 
 ## 0.21.56