+++ packages/components/FocusTrapZone/CHANGELOG.md
@@ -1,8 +1,37 @@
 # Change Log - @fluentui-react-native/focus-trap-zone
 
-This log was last generated on Mon, 30 Jan 2023 07:12:38 GMT and should not be manually modified.
+This log was last generated on Fri, 03 Feb 2023 01:49:36 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.58
+
+Fri, 03 Feb 2023 01:49:36 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.17
+
+## 0.9.57
+
+Thu, 02 Feb 2023 01:29:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.16
+
+## 0.9.56
+
+Wed, 01 Feb 2023 22:33:15 GMT
+
+### Patches
+
+- Enable @typescript-eslint/consistent-type-imports (30809111+acoates-ms@users.noreply.github.com)
+- Bump @fluentui-react-native/adapters to v0.10.1
+- Bump @fluentui-react-native/component-cache to v1.4.3
+- Bump @fluentui-react-native/interactive-hooks to v0.22.15
+- Bump @uifabricshared/foundation-composable to v0.11.2
+- Bump @uifabricshared/foundation-settings to v0.12.2
 
 ## 0.9.55