+++ packages/experimental/Link/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/experimental-link
 
-This log was last generated on Mon, 30 Jan 2023 07:12:38 GMT and should not be manually modified.
+This log was last generated on Fri, 03 Feb 2023 01:49:36 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.23
+
+Fri, 03 Feb 2023 01:49:36 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/link to v0.18.14
+
+## 0.4.22
+
+Thu, 02 Feb 2023 01:29:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/link to v0.18.13
+
+## 0.4.21
+
+Wed, 01 Feb 2023 22:33:17 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/link to v0.18.12
 
 ## 0.4.20